refactor: move connector leaves (gitea, oscqam, openqa/) to data_sources/
Step 7, part 1 of 3: the no-split moves. `git mv` only --- function
bodies and class definitions are unchanged.

Moves:
  mtui/connector/gitea.py    -> mtui/data_sources/gitea.py
  mtui/connector/oscqam.py   -> mtui/data_sources/oscqam.py
  mtui/connector/openqa/     -> mtui/data_sources/openqa/

Importer rewrites (callers of those modules):
  mtui/commands/approve.py, apicall.py: `from ..connector import OSC, Gitea`
    -> `from ..data_sources import OSC, Gitea`
  mtui/commands/simpleset.py, reloadoqa.py, mtui/types/updateid.py:
    `from ..connector.openqa import KernelOpenQA`
    -> `from ..data_sources.openqa import KernelOpenQA`
  mtui/test_reports/sl_report.py:
    `from ..connector.gitea import Gitea`
    -> `from ..data_sources.gitea import Gitea`

mtui/data_sources/__init__.py re-exports OSC and Gitea to match the
old connector/__init__.py surface for those names. connector/__init__.py
is downgraded to a transitional shim re-exporting only the two names
still living under connector/ (QEMIncident, DashboardAutoOpenQA);
those move in the next two commits (qem_dashboard split, oqa_search
split).

Test edits: tests/test_gitea.py, test_oscqam.py (12 `mock.patch()`
strings rewritten), test_openqa_connector.py (1 import + 1 patch
string). Logger strings (`getLogger("mtui.connector.oscqam")` etc.)
are unchanged --- consistent with the convention established in the
support/ and cli/ moves: logger names are external contracts
independent of the import path.

`mtui/types/oqaresults.py` docstring updated for the openqa class
references (cosmetic).

Gates: ruff format/check + ty check + pytest (`not slow and not
network`) all green --- 989 passed.
