+++ mtui/commands/apicall.py
@@ -13,7 +13,7 @@ from typing import ClassVar, final
 
 from ..cli.argparse import ArgumentParser
 from ..cli.completion import complete_choices
-from ..connector import OSC, Gitea
+from ..data_sources import OSC, Gitea
 from ..support.exceptions import GiteaError
 from ..support.misc import requires_update
 from ..types import RequestKind
