+++ mtui/commands/reloadoqa.py
@@ -2,8 +2,8 @@
 
 from logging import getLogger
 
-from ..connector.openqa import KernelOpenQA
 from ..connector.qem_dashboard import DashboardAutoOpenQA
+from ..data_sources.openqa import KernelOpenQA
 from ..support.misc import requires_update
 from . import Command
