+++ mtui/data_sources/__init__.py
@@ -1,1 +1,6 @@
 """External data-source clients (Gitea, OSC, openQA, QEM Dashboard)."""
+
+from .gitea import Gitea
+from .oscqam import OSC
+
+__all__ = ["OSC", "Gitea"]
