+++ tests/test_oscqam.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import pytest
 
-from mtui.connector.oscqam import OSC
+from mtui.data_sources.oscqam import OSC
 from mtui.types import RequestKind
 
 
@@ -22,7 +22,7 @@ class TestOSCInit:
 
 
 class TestOSCOperations:
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_approve(self, mock_check_call, osc):
         """Test approve builds correct command."""
         osc.approve(["qam-sle"])
@@ -34,7 +34,7 @@ class TestOSCOperations:
         assert "-G" in cmd
         assert "qam-sle" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_assign(self, mock_check_call, osc):
         """Test assign builds correct command."""
         osc.assign(["qam-sle"])
@@ -42,7 +42,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "assign" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_unassign(self, mock_check_call, osc):
         """Test unassign builds correct command."""
         osc.unassign(["qam-sle"])
@@ -50,7 +50,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "unassign" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_reject_with_reason_and_message(self, mock_check_call, osc):
         """Test reject includes reason and message."""
         osc.reject(["qam-sle"], "bug found", "details here")
@@ -61,7 +61,7 @@ class TestOSCOperations:
         assert "bug found" in cmd
         assert "-M" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_comment(self, mock_check_call, osc):
         """Test comment builds correct command."""
         osc.comment("test comment")
@@ -69,7 +69,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "comment" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_multiple_groups(self, mock_check_call, osc):
         """Test command with multiple groups adds -G for each."""
         osc.approve(["qam-sle", "qam-kernel"])
@@ -78,7 +78,7 @@ class TestOSCOperations:
         g_indices = [i for i, x in enumerate(cmd) if x == "-G"]
         assert len(g_indices) == 2
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_empty_groups(self, mock_check_call, osc):
         """Test command with no groups omits -G."""
         osc.comment("hello")
@@ -86,7 +86,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "-G" not in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_skip_template_for_pi(self, mock_check_call, osc):
         """Test --skip-template is added for PI kind."""
         osc.rrid.kind = RequestKind.PI
@@ -95,7 +95,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "--skip-template" in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_no_skip_template_for_sle(self, mock_check_call, osc):
         """Test --skip-template is NOT added for non-PI/non-SLFO kinds."""
         osc.rrid.kind = RequestKind.MAINTENANCE
@@ -104,7 +104,7 @@ class TestOSCOperations:
         cmd = mock_check_call.call_args[0][0]
         assert "--skip-template" not in cmd
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_calledprocesserror_logged(self, mock_check_call, osc):
         """Test CalledProcessError is caught and logged."""
         from subprocess import CalledProcessError
@@ -112,13 +112,13 @@ class TestOSCOperations:
         mock_check_call.side_effect = CalledProcessError(1, "osc")
         osc.approve(["qam-sle"])  # should not raise
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_filenotfounderror_logged(self, mock_check_call, osc):
         """Test FileNotFoundError is caught and logged."""
         mock_check_call.side_effect = FileNotFoundError("osc not found")
         osc.approve(["qam-sle"])  # should not raise
 
-    @patch("mtui.connector.oscqam.check_call")
+    @patch("mtui.data_sources.oscqam.check_call")
     def test_api_url(self, mock_check_call, osc):
         """Test API URL is passed correctly."""
         osc.approve(["qam-sle"])
